@willtebbutt 2022-08-19
@theogf let me know whether my explanation of the testing is sufficient, and I'll add a docstring + merge